fix: postのtypesファイルを分離
