+++ features/types/index.ts
@@ -1,0 +1,10 @@
+export type PostProps = {
+  id: number;
+  title: string;
+  // author: {
+  //   name: string;
+  // };
+  content: string;
+  published: boolean;
+  createdAt: string;
+};
